[refactor]: hid sensitive data, moved docker-compose and DaysOfWeek files to other directories, docs
