+++ src/main/java/org/example/crmedu/application/controller/package-info.java
@@ -1,4 +1,10 @@
 /**
- * This package contains all controllers of this application and advicer for central processing of HTTP responses.
+ * Provides classes responsible for managing data at the transport layer and centralizing exception handling.
+ * <p>
+ * The {@code AdviceController} class provides handlers for exceptions that may arise in the system. These exceptions can occur while working with data at the
+ * persistence or business layer, as well as during validation at the transport layer.
+ * </p>
+ * Other classes, such as {@code OrganizationController}, {@code SubjectController}, {@code TutorController}, and {@code UserController}, provide RESTful
+ * endpoints for data management.
  */
 package org.example.crmedu.application.controller;
