+++ src/main/java/org/example/crmedu/domain/model/TutorSchedule.java
@@ -3,6 +3,7 @@ package org.example.crmedu.domain.model;
 import java.time.OffsetTime;
 import lombok.Data;
 import lombok.experimental.Accessors;
+import org.example.crmedu.domain.enums.DaysOfWeek;
 
 /**
  * A class representing a tutor schedule entity. This class is used as a domain model. Contains information about tutor's work time.
