+++ src/main/resources/application.yaml
@@ -7,9 +7,9 @@ spring:
     liquibase-schema: crmedu
     change-log: classpath:db/changelog/db.changelog-master.yaml
   datasource:
-    url: jdbc:postgresql://localhost:5435/postgres
-    username: postgres
-    password: changeme
+    username: ${DATABASE_USERNAME}
+    password: ${DATABASE_PASSWORD}
+    url: ${DATABASE_URL}
     driver-class-name: org.postgresql.Driver
   jpa:
     properties:
